requirements.txt
django
psycopg2
djangorestframework
requests
django-filter